refactor: migrate to TanStack Start SPA
