--- index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <title>Vite React Chakra Starter</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <script src="/src/main.tsx" type="module"></script>
-  </body>
-</html>
