+++ knip.ts
@@ -1,7 +1,7 @@
 import type { KnipConfig } from 'knip';
 
 const config: KnipConfig = {
-  entry: ['src/main.tsx'],
+  entry: ['src/router.tsx'],
   ignore: ['src/**/*.gen.ts'],
   ignoreBinaries: ['changelogithub'],
   project: ['src/**/*.{ts,tsx,js,jsx,css,scss}'],
