+++ package.json
@@ -45,6 +45,7 @@
     "@tanstack/react-query-devtools": "^5.101.4",
     "@tanstack/react-router": "^1.170.18",
     "@tanstack/react-router-devtools": "^1.167.0",
+    "@tanstack/react-start": "1.168.31",
     "next-themes": "^0.4.6",
     "react": "^19.2.8",
     "react-dom": "^19.2.8",
@@ -60,7 +61,6 @@
     "@julr/vite-plugin-validate-env": "^2.2.2",
     "@rolldown/plugin-babel": "^0.2.3",
     "@tanstack/devtools-vite": "^0.8.3",
-    "@tanstack/router-plugin": "^1.168.23",
     "@types/node": "^26.1.2",
     "@types/react": "19.2.17",
     "@types/react-dom": "^19.2.3",
