+++ src/lib/layout/index.tsx
@@ -3,7 +3,6 @@ import type { ReactNode } from 'react';
 
 import { Footer } from './components/footer';
 import { Header } from './components/header';
-import { Meta } from './components/meta';
 
 type LayoutProps = {
   children: ReactNode;
@@ -11,7 +10,6 @@ type LayoutProps = {
 
 export const Layout = ({ children }: LayoutProps) => (
   <Box margin="0 auto" maxWidth={800} transition="0.5s ease-out">
-    <Meta />
     <Flex margin="8" minHeight="90vh" wrap="wrap">
       <Header />
       <Box as="main" marginY={22} width="full">
