--- src/main.tsx
@@ -1,53 +1,0 @@
-import { QueryClientProvider } from '@tanstack/react-query';
-import { createRouter, RouterProvider } from '@tanstack/react-router';
-import { StrictMode } from 'react';
-import ReactDOM from 'react-dom/client';
-
-import { Provider } from '@/lib/components/ui/provider';
-import Page404 from '@/lib/pages/404';
-import { queryClient } from '@/lib/services/constants';
-
-// Import the generated route tree
-import { routeTree } from './routeTree.gen';
-
-// fonts
-import '@fontsource-variable/plus-jakarta-sans';
-
-// Create a new router instance
-const router = createRouter({
-  context: {
-    queryClient,
-  },
-  defaultNotFoundComponent: () => <Page404 />,
-  defaultPendingComponent: () => (
-    <div className="mx-auto">
-      <p>Loading...</p>
-    </div>
-  ),
-  defaultPreload: 'intent',
-  defaultStructuralSharing: true,
-  routeTree,
-  scrollRestoration: true,
-});
-
-// Register the router instance for type safety
-declare module '@tanstack/react-router' {
-  interface Register {
-    router: typeof router;
-  }
-}
-
-// Render the app
-const rootElement = document.getElementById('app');
-if (rootElement && !rootElement.innerHTML) {
-  const root = ReactDOM.createRoot(rootElement);
-  root.render(
-    <StrictMode>
-      <Provider>
-        <QueryClientProvider client={queryClient}>
-          <RouterProvider router={router} />
-        </QueryClientProvider>
-      </Provider>
-    </StrictMode>
-  );
-}
